Scrape hourly
Up until now, scraping was scheduled based on
the server response. The server response contained
a timestamp, which indicated when will the server
publish new rankings.

However, the server response was a little inconsistent
and sometimes this timestamp differed by a couple of
seconds, resulting in multiple scrapes.

Other changes:
  * Moving to JDK 17 with Vertx 4.2.2
  * Bumping Log4J to 2.16.0